pyproject.toml
[project]
name = "sdm-project"
version = "0.1.0"
description = "Academic Property Graph – SDM Lab"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lxml>=6.0.2",
    "neo4j>=6.1.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
]
